requirements.txt
asn1crypto==0.22.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cryptography==2.0.3
decorator==4.1.2
enum34==1.1.6
functools32==3.2.3.post2
gitdb2==2.0.3
GitPython==2.1.7
idna==2.6
ipaddress==1.0.18
ipdb==0.10.3
ipython==5.4.1
ipython-genutils==0.2.0
Jinja2==2.9.6
jsonschema==2.6.0
MarkupSafe==1.0
paramiko==2.2.1
pathlib==1.0.1
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
progress==1.3
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.3.3
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.1.2
python-dateutil==2.6.1
PyYAML==3.12
pyzmq==16.0.2
requests==2.18.4
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
smmap2==2.0.3
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
zmq==0.0.0
